+++ app/src/main/java/com/vrcareer/b4b/model/WithdrawalRequest.kt
@@ -2,7 +2,7 @@ package com.vrcareer.b4b.model
 
 data class WithdrawalRequest(
     var id: String? = null,
-    var amount: Int? = null,
+    var amount: Float? = null,
     var user_id: String? = null,
     var time_of_request: Long? = null,
     var time_of_approval: Long? = null,
